+++ packages/workflows/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @questpie/workflows
 
+## 3.28.10
+
+### Patch Changes
+
+- Updated dependencies []:
+  - questpie@3.28.10
+  - @questpie/admin@3.28.10
+
 ## 3.28.9
 
 ### Patch Changes